Initial commit
Add the files necessary for a MEAN.JS style application.  This is
mostly just scaffolding, with a few files useful in development
(logger.js for example).  The project is setup to use “meanjs-template”
as the name, so a find/replace for “meanjs” or “MEAN” should be done
when using this template.

Some tests have been written to illustrate how server and client side
tests can be written.  In addition, most of the build steps have been
setup in the Gruntfile.js.

On the client side, a main “app” view has been created in the “global”
module, and two example views have been created in the “home” module,
one that wraps the other.