Autoprefixer
Use the grunt postcss task to run autoprefixer on our generated CSS.
Create a grunt task to incorporate all the css generation tasks into
one task (so this will generate the CSS from LESS, then run the
autoprefixer to prefix the CSS).

Also include this generate-css task within the test tasks, just to be
sure (more for protractor).  Formatting of the package.json is the
reason for the huge diff.