Enable JUnit reports for CI
To allow the tests to run in a way that the continuous integration
environment (Jenkins) can monitor the tests, enable JUnit reporting for
all tests (server side, client side, and end to end).  Output all the
tests to the `test-reports` directory.

Update `jasmine-node`, `karma`, and `protractor` to use JUnit reporter
when the CI flag is enabled.  Add documentation on how to run the tests
in continuous integration mode.