Use Git revision on production assets
As a continuation on the Git revision use, use the (short) Git revision
on the production asset filenames to both state the version used to
generate, but also to cache-bust when we release changes (the browser
will initiate a new request since the name of the file has changed).
Granted the cache can, and probably should, be setup through some web
server such as nginx, this at least prepares us to handle that case.

Also switch to using `git-rev-sync` to make processing of the Git
revision easier.