Replace logger, npm start changes
To avoid generating the stack trace to find the namespace for our
logger, use the `debug-caller` library which does the work for us.

Default the `npm start` script to use the development environment. This
is to avoid having to type that in every time, since it’s pretty much
the one environment used for that command.  Document the change in the
readme.

Also some minor changes to remove an unused Angular app config parameter,
express config changes, bower resolutions, and rename the e2e test to a
more generic name.